[RD2] Consolidate JSON mock data for recurring donation object describes and getRecurringData
