--- force-app/main/default/lwc/rd2EntryFormDonorSection/__tests__/data/recurringDataAccountResponse.json
@@ -1,6 +1,0 @@
-{
-  "DonorType": "Account",
-  "Period": "Monthly",
-  "Frequency": 1,
-  "RecurringType": "Open"
-}
