--- force-app/main/default/lwc/rd2EntryFormDonorSection/__tests__/data/recurringDataContactResponse.json
@@ -1,6 +1,0 @@
-{
-  "DonorType": "Contact",
-  "Period": "Monthly",
-  "Frequency": 1,
-  "RecurringType": "Open"
-}
